% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type2 data.R
\name{data_update2}
\alias{data_update2}
\title{generate lottodata2}
\usage{
data_update2(data = data_update())
}
\value{
lotto data2
}
\description{
generate lottodata2
}
\examples{
data_update2()
round(cor(data2),3)
}
