% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/del.R
\name{del}
\alias{del}
\title{delete of number}
\usage{
del(n = 100, tol = 1, keep = NULL, drop = NULL)
}
\arguments{
\item{n}{= Number to consider.}

\item{tol}{=Range of absolute values}
}
\description{
delete of number
}
