% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample type1.R
\name{lotto1}
\alias{lotto1}
\title{sample type1}
\usage{
lotto1(n, keep = NULL, drop = NULL, hold = NULL, a = NULL, write = F,
  wd = NULL)
}
\arguments{
\item{n}{= number of sample}
}
\description{
sample type1
}
\examples{
lotto1(5)
(k<-sort(num_per()[as.numeric(data[1,1:6])],dec=T))
lotto1(n=5,keep=k[1])
lotto1(n=5000,hold=as.numeric(names(k[1])))
}
