% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample type2.R
\name{lotto2}
\alias{lotto2}
\title{sample2}
\usage{
lotto2(n, count = 1, keep = NULL, drop = NULL, hold = NULL, up = NULL,
  header = 10, write = F)
}
\arguments{
\item{n}{= The number of times the sample is pulled}

\item{keep}{= The number to keep the probability of}

\item{drop}{=The number corresponding to 'drop' does not appear.}

\item{hold}{=This number appears unconditionally.}
}
\description{
sample2
}
\examples{
(k<-sort(num_per()[as.numeric(data[1,1:6])],dec=T))
lotto2(n=100,hold=as.numeric(names(k[1])),header=10)
}
