% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample type3.R
\name{lotto3}
\alias{lotto3}
\title{lotto}
\usage{
lotto3(n, a = NULL, b = NULL, head = 10, write = F)
}
\arguments{
\item{n}{= The number of times the sample is pulled}
}
\description{
lotto
}
\examples{
lotto3(5)
}
