% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate lottodata.R
\name{numdata}
\alias{numdata}
\title{generate lottodata}
\usage{
numdata(first = 1, last, num = list(), output = list(), k = 1,
  update = F, sleep = 1)
}
\description{
generate lottodata
}
\examples{
url<-'https://search.naver.com/search.naver?sm=tab_hty.top&where=nexearch&query=\%EB\%A1\%9C\%EB\%98\%90&oquery=\%EB\%A1\%9C\%EB\%98\%90&tqi=TmrKidpVuFdsssc0EvVssssssUd-075322'
line<-read_html(url,encoding="UTF-8")
p1<-html_nodes(line,css='._lotto-btn-current em')\%>\%html_text()
last=as.numeric(substr(p1,1,3))
(data<-numdata(last=last,update=T,sleep=0))

}
